Implement automation of some build tasks using Gradle
