+++ build.gradle
@@ -2,6 +2,7 @@ plugins {
     id 'java'
     id 'application'
     id 'com.github.johnrengelman.shadow' version '7.1.2'
+    id 'checkstyle'
 }
 
 repositories {
@@ -28,14 +29,19 @@ test {
 }
 
 application {
-    mainClass.set("seedu.duke.Duke")
+    mainClass.set("strand.Strand")
 }
 
 shadowJar {
-    archiveBaseName = "duke"
+    archiveBaseName = "strand"
     archiveClassifier = null
 }
 
 run{
     standardInput = System.in
+//    enableAssertions = true
+}
+
+checkstyle {
+    toolVersion = '10.2'
 }
