--- src/main/java/strand/Exceptions/StrandDescNotFoundException.java
@@ -1,15 +1,0 @@
-package strand.Exceptions;
-
-public class StrandDescNotFoundException extends StrandException {
-    protected String part;
-
-    public StrandDescNotFoundException(String s) {
-        this.part = s;
-    }
-
-    @Override
-    public String toString() {
-        return this.part + " not found " + super.toString()
-                + "\nPlease include a " + this.part + " for this task";
-    }
-}
