--- src/main/java/strand/Exceptions/StrandException.java
@@ -1,8 +1,0 @@
-package strand.Exceptions;
-
-public abstract class StrandException extends Exception {
-    @Override
-    public String toString() {
-        return "(×_×;）";
-    }
-}
