--- src/main/java/strand/Exceptions/StrandFileNotFoundException.java
@@ -1,8 +1,0 @@
-package strand.Exceptions;
-
-public class StrandFileNotFoundException extends StrandException {
-    @Override
-    public String toString() {
-        return "File not found " + super.toString();
-    }
-}
