--- src/main/java/strand/Exceptions/StrandWrongCommandException.java
@@ -1,8 +1,0 @@
-package strand.Exceptions;
-
-public class StrandWrongCommandException extends StrandException {
-    @Override
-    public String toString() {
-        return "Command not found " + super.toString();
-    }
-}
