+++ src/main/java/strand/exception/StandInvalidDateException.java
@@ -1,5 +1,8 @@
-package strand.Exceptions;
+package strand.exception;
 
+/**
+ * Exception thrown when the date given is invalid.
+ */
 public class StandInvalidDateException extends StrandException {
     protected String date;
 
@@ -8,8 +11,8 @@ public class StandInvalidDateException extends StrandException {
     }
 
     @Override
-    public String toString() {
-        return this.date + " is not a valid date " + super.toString()
+    public String getMessage() {
+        return this.date + " is not a valid date " + super.getMessage()
                 + "\nPlease include a date of the correct format (e.g. 2/12/2019 1800)";
     }
 }
