+++ src/main/java/strand/exception/StrandException.java
@@ -1,0 +1,11 @@
+package strand.exception;
+
+/**
+ * Base class for all exceptions in the Strand application.
+ */
+public abstract class StrandException extends Exception {
+    @Override
+    public String getMessage() {
+        return "(×_×;）";
+    }
+}
