+++ src/main/java/strand/exception/StrandWrongCommandException.java
@@ -1,0 +1,11 @@
+package strand.exception;
+
+/**
+ * Thrown when an unrecognized command is encountered.
+ */
+public class StrandWrongCommandException extends StrandException {
+    @Override
+    public String getMessage() {
+        return "Command not found " + super.getMessage();
+    }
+}
