+++ src/main/java/strand/task/Event.java
@@ -1,8 +1,8 @@
-package strand.Tasks;
-
-import strand.Exceptions.StrandException;
+package strand.task;
 
 import java.time.LocalDateTime;
+
+import strand.exception.StrandException;
 
 /**
  * The strand.Tasks.Event class represents a task that occurs within a specific time range.
