+++ src/main/java/strand/task/Task.java
@@ -1,8 +1,4 @@
-package strand.Tasks;
-
-import strand.Exceptions.StandInvalidDateException;
-import strand.Exceptions.StrandDescNotFoundException;
-import strand.Exceptions.StrandException;
+package strand.task;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
@@ -10,8 +6,17 @@ import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeFormatterBuilder;
 import java.time.format.DateTimeParseException;
 
+import strand.exception.StandInvalidDateException;
+import strand.exception.StrandDescNotFoundException;
+import strand.exception.StrandException;
+
 /**
- * The strand.Tasks.Task class is an abstract representation of a task
+ * The {@code Task} class is an abstract representation of a task.
+ * <p>
+ * This class provides common properties and methods for handling tasks, including
+ * description management, status tracking (whether the task is done or not),
+ * and date parsing/formatting.
+ * </p>
  */
 public abstract class Task {
     protected String description;
@@ -30,10 +35,11 @@ public abstract class Task {
     protected DateTimeFormatter formatterDisplay = DateTimeFormatter.ofPattern("MMM dd yyyy, HH:mm");
 
     /**
-     * Constructs a new strand.Tasks.Task with the specified description.
+     * Constructs a new {@code Task} with the specified description.
      * By default, the task is not marked as done.
      *
      * @param description The description of the task.
+     * @throws StrandException if the description is empty.
      */
     public Task(String description) throws StrandException {
         if (description.isEmpty()) {
@@ -43,6 +49,19 @@ public abstract class Task {
         this.isDone = false;
     }
 
+    /**
+     * Parses a date string and returns a {@code LocalDateTime} object.
+     * <p>
+     * The method tries to parse the date using multiple formats, both with and
+     * without time information. If parsing as {@code LocalDateTime} fails, it
+     * attempts to parse as {@code LocalDate} and returns the start of the day
+     * as the {@code LocalDateTime}.
+     * </p>
+     *
+     * @param date The date string to be parsed.
+     * @return The parsed {@code LocalDateTime} object.
+     * @throws StrandException if the date cannot be parsed.
+     */
     public LocalDateTime parseDate(String date) throws StrandException {
         try {
             return LocalDateTime.parse(date, formatterWithTime);
@@ -57,20 +76,39 @@ public abstract class Task {
         }
     }
 
+    /**
+     * Formats a {@code LocalDateTime} object into a string for display purposes.
+     *
+     * @param date The {@code LocalDateTime} object to be formatted.
+     * @return The formatted date string.
+     */
     public String parseOutputDate(LocalDateTime date) {
         return date.format(formatterDisplay);
     }
 
+    /**
+     * Returns the status of the task as a string, indicating whether it is done.
+     *
+     * @return A string representing the status of the task.
+     */
     private String getStatus() {
         return "[" + (isDone ? "X" : " ") + "] ";
     }
 
+    /**
+     * Returns the type of the task.
+     * <p>
+     * This method is abstract and must be implemented by subclasses.
+     * </p>
+     *
+     * @return A string representing the type of the task.
+     */
     abstract String getType();
 
     /**
-     * Returns the string representation of the task
+     * Returns the string representation of the task, including its status and description.
      *
-     * @return A string representing the task
+     * @return A string representing the task.
      */
     @Override
     public String toString() {
@@ -78,21 +116,26 @@ public abstract class Task {
     }
 
     /**
-     * Marks the task as done
+     * Marks the task as done.
      */
     public void markAsDone() {
         this.isDone = true;
     }
 
     /**
-     * Marks the task as not done
+     * Marks the task as not done.
      */
     public void markAsNotDone() {
         this.isDone = false;
     }
 
+    /**
+     * Returns the string representation of the task for file storage,
+     * including its status and description.
+     *
+     * @return A string formatted for file storage.
+     */
     public String getFile() {
         return String.format("%s | %s", this.isDone ? "1" : "0", this.description);
     }
-
 }
